requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
decorator==4.4.2
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
pycparser==2.20
python-telegram-bot==12.4.2
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.0
